Don't clear any tstprefs() or tstlocalrules() settings with clear_localrules()
git-svn-id: https://svn.apache.org/repos/asf/spamassassin/trunk@1901348 13f79535-47bb-0310-9956-ffa450edef68

+++ t/SATest.pm
@@ -284,6 +284,9 @@ sub clear_localrules {
     # Keep some useful, should not contain any rules
     next if $file =~ /10_default_prefs.cf$/;
     next if $file =~ /20_aux_tlds.cf$/;
+    # Keep our own tstprefs() or tstlocalrules()
+    next if $file =~ /99_test_prefs.cf$/;
+    next if $file =~ /99_test_rules.cf$/;
     unlink $file;
   }
 }
